# swagger/swagger.yaml
openapi: 3.0.0
info:
  title: RareCare - Maladies API
  version: 1.0.0
servers:
  - url: http://localhost:8001/api
paths:
  /maladies:
    get:
      summary: Get all maladies
      tags:
        - Maladies
      responses:
        '200':
          description: List of maladies
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Maladie'
    post:
      summary: Create a new maladie
      tags:
        - Maladies
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MaladieInput'
      responses:
        '201':
          description: Maladie created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Maladie'
        '422':
          description: Validation error
  /maladies/{id}:
    get:
      summary: Get a single maladie
      tags:
        - Maladies
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Maladie details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Maladie'
        '404':
          description: Maladie not found
    put:
      summary: Update a maladie
      tags:
        - Maladies
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MaladieInput'
      responses:
        '200':
          description: Maladie updated
        '404':
          description: Maladie not found
    delete:
      summary: Delete a maladie
      tags:
        - Maladies
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Maladie deleted
        '404':
          description: Maladie not found

components:
  schemas:
    Maladie:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        rare:
          type: boolean
        created_at:
          type: string
        updated_at:
          type: string
    MaladieInput:
      type: object
      properties:
        name:
          type: string
        rare:
          type: boolean
      required:
        - name